agent2/requirements.txt
anthropic
python-dotenv
gspread
google-auth
fastapi
uvicorn[standard]
pytest
pytest-mock
